mythicalnuc/MythicalVisionEdge/modules/RTSPImageCapturePY/requirements.txt
imutils==0.5.4
Pillow==9.4.0
azure-storage-blob
numpy
opencv-python
azure-iot-device~=2.7.0
